swe_flutterrr/Rest API/requirements.txt
autopep8==1.5.6
click==7.1.2
cycler==0.10.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
Werkzeug==1.0.1
